New folder/otp_auth_project/requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
gunicorn
dj-database-url
whitenoise
psycopg2-binary
